Fix axios refresh token

+++ src/app/AxiosClient.ts
@@ -1,4 +1,5 @@
-import axios, { AxiosRequestHeaders } from 'axios';
+import axios, { AxiosError, AxiosRequestHeaders } from 'axios';
+import Cookies from 'js-cookie';
 
 import { API_URL } from '@/config/index';
 import jwtService from '@/services/JwtService';
@@ -59,8 +60,7 @@ function createAxiosResponseInterceptor(): void {
       // Reject if the error message is not 'Unauthorized'
       if (
         error.response.data.message !== 'Unauthorized' ||
-        config.url === '/auth/refresh' ||
-        !jwtService.getToken()
+        config.url === '/auth/refresh'
       ) {
         return Promise.reject(error);
       }
@@ -84,10 +84,13 @@ function createAxiosResponseInterceptor(): void {
           config.headers.Authorization = `Bearer ${refreshResponse.access_token}`;
           return axiosClient(config);
         })
-        .catch((refreshError) => {
-          // Remove the access token and redirect to sign in page
-          jwtService.removeToken();
-          window.location.replace('/sign-in?error=expired');
+        .catch((refreshError: AxiosError) => {
+          if (refreshError?.response?.status === 401) {
+            // Remove the access token and redirect to sign in page
+            Cookies.remove('Refresh');
+            jwtService.removeToken();
+            window.location.replace('/sign-in?error=expired');
+          }
           return Promise.reject(refreshError);
         })
         .finally(createAxiosResponseInterceptor);
